🩹 Fixing empty state on initial pop
