+++ utils/classList.js
@@ -1,0 +1,77 @@
+/******************************************************************************
+**	@Author:				Rarity Extended
+**	@Twitter:				@MajorTom_eth
+**	@Date:					Saturday September 11th 2021
+**	@Filename:				ClassList.js
+******************************************************************************/
+
+const	classes = {
+	'Barbarian': {
+		name: 'Barbarian',
+		img: '/barbarian.png',
+		description: 'No friend of the books, unlike any librarian.\nStrength, weapons, and anger serve the Barbarian',
+		id: 1,
+	},
+	'Bard': {
+		name: 'Bard',
+		img: '/bard.png',
+		description: 'Words, songs, and music are certainly not hard.\n The magic of the voice is the weapon of the Bard',
+		id: 2,
+	},
+	'Cleric': {
+		name: 'Cleric',
+		img: '/cleric.png',
+		description: 'In the world of adventure, pains and wounds are quite generic.\nIf you live a life or danger, you best know a Cleric',
+		id: 3,
+	},
+	'Druid': {
+		name: 'Druid',
+		img: '/druid.png',
+		description: 'All life is connected in something that is rather fluid.\nThe trees, insects, and animals are all friends of the Druid',
+		id: 4,
+	},
+	'Fighter': {
+		name: 'Fighter',
+		img: '/fighter.png',
+		description: 'Scorn should not be directed at one with a dream to be a writer.\nBut tactics and sword play are what drive the Fighter',
+		id: 5,
+	},
+	'Monk': {
+		name: 'Monk',
+		img: '/monk.png',
+		description: 'Some pursue vanity, and others just want to get drunk.\nInner peace, and control of the body are the goals of a monk',
+		id: 6,
+	},
+	'Paladin': {
+		name: 'Paladin',
+		img: '/paladin.png',
+		description: 'Some hearts when inspected are found with malice therein.\nBut righteous and honor are the tenets of the Paladin',
+		id: 7,
+	},
+	'Ranger': {
+		name: 'Ranger',
+		img: '/ranger.png',
+		description: 'Most, avoid, flee, and fear only a little bit of danger.\nWith a bow in the wilderness, you might find a Ranger',
+		id: 8,
+	},
+	'Rogue': {
+		name: 'Rogue',
+		img: '/rogue.png',
+		description: 'The rich are rich and the poor are poor is in vogue.\nBut with sticky fingers and sharp daggers you find the Rogue',
+		id: 9,
+	},
+	'Sorcerer': {
+		name: 'Sorcerer',
+		img: '/sorcerer.png',
+		description: 'A scholarly teacher of magic seems like a torturer.\nBut this is not of concern to the innate magic of a Sorcerer',
+		id: 10,
+	},
+	'Wizard': {
+		name: 'Wizard',
+		img: '/wizard.png',
+		description: 'Many waste their time on a log, idle like a lizard.\nHowever, through study, immense power is granted to the Wizard',
+		id: 11,
+	},
+};
+
+export default classes;
